process-user-data: move logic in pkgs
- Move the logic into pkgs
- Use the log module for log output
