--- cmd/process-user-data/types.go
@@ -1,29 +1,0 @@
-package main
-
-const (
-	programName   = "process-user-data"
-	providerAzure = "azure"
-	providerAws   = "aws"
-
-	defaultAgentConfigPath  = "/etc/agent-config.toml"
-	defaultAuthJsonFilePath = "/run/peerpod/auth.json"
-)
-
-type Config struct {
-	daemonConfigPath     string
-	agentConfigPath      string
-	userDataFetchTimeout int
-	authJsonPath         string
-}
-
-type Endpoints struct {
-	Allowed []string `toml:"allowed"`
-}
-
-type AgentConfig struct {
-	EnableSignatureVerification bool      `toml:"enable_signature_verification"`
-	ServerAddr                  string    `toml:"server_addr"`
-	AaKbcParams                 string    `toml:"aa_kbc_params"`
-	ImageRegistryAuthFile       string    `toml:"image_registry_auth_file"`
-	Endpoints                   Endpoints `toml:"endpoints"`
-}
